Implement auto-select functionality across conversation components
- Added AUTO_SELECT_ENABLED configuration to manage auto-select behavior.
- Updated ChatContextProgress to conditionally render based on auto-select state.
- Enhanced ConversationAccordion to utilize auto-select logic for conversation item interactions.
- Refactored NavigationButton and ConversationLinks to support dynamic border colors based on auto-select status.
- Replaced AutoSelectConversationItem with AutoSelectConversations for improved integration in the conversation flow.
