+++ echo/frontend/src/components/conversation/ConversationLinks.tsx
@@ -4,8 +4,10 @@ import { I18nLink } from "@/components/common/i18nLink";
 
 export const ConversationLinks = ({
   conversations,
+  color,
 }: {
   conversations: Conversation[];
+  color?: string;
 }) => {
   const { projectId } = useParams();
 
@@ -16,7 +18,7 @@ export const ConversationLinks = ({
           key={conversation.id}
           to={`/projects/${projectId}/conversation/${conversation.id}/overview`}
         >
-          <Anchor size="xs">{conversation.participant_name}</Anchor>
+          <Anchor size="xs" c={color}>{conversation.participant_name}</Anchor>
         </I18nLink>
       )) ?? null}
     </Group>
